+++ src/kenze/__init__.py
@@ -38,7 +38,7 @@ from .ops import (
     validate,
 )
 
-__version__ = "0.7.0"
+__version__ = "0.7.1"
 
 
 def sift(input, output, con=None, quiet=True, **steps):